feat: Adding Embed Class & Renaming files.
