+++ src/lib/util/valeriyya.types.ts
@@ -19,7 +19,7 @@ export type AbstractConstructor<T> = abstract new (...args: any[]) => T;
 export type Nullish = null | undefined;
 
 export interface ICommandInteraction extends CommandInteraction {
-
+    // db: GuildDb 
 }
 export type ICommandExecute = (interaction: CommandInteraction) => Promise<string | MessagePayload | InteractionReplyOptions | void> | InteractionReplyOptions | string| void;
